fixed #130 and varies map issues

+++ EatNow/Extensins/ENUtil.m
@@ -91,9 +91,9 @@ void ENLogError(NSString *fmt,...){
 	}else if (hours >= 1){
 		timeStr = [NSString stringWithFormat:@"%.1f hours", hours + minutes/60];
 	}else if(minutes >= 1){
-		timeStr = [NSString stringWithFormat:@"%ld minutes",(long)minutes];
+		timeStr = [NSString stringWithFormat:@"%.0f minutes",minutes];
 	}else{
-		timeStr = [NSString stringWithFormat:@"%ld seconds",(long)seconds];
+		timeStr = [NSString stringWithFormat:@"%.0f seconds",seconds];
 	}
 	return timeStr;
 }
